requirements.txt
# Основные зависимости
torch>=2.0.1
transformers>=4.30.2
networkx>=3.1
python-dotenv>=1.0.0
numpy>=1.23.5

# Дополнительные модули
requests>=2.28.2  # Для веб-запросов
pandas>=1.5.3  # Для анализа данных
tqdm>=4.65.0  # Индикаторы прогресса

# Разработка
pytest>=7.2.2  # Тестирование
black>=23.3.0  # Форматирование кода